chore: renamed classes to use singular signalering instead of signaleringen
