+++ src/main/kotlin/net/atos/zac/app/signalering/model/RESTSignaleringInstellingen.kt
